BLD: do not create broken symlink

+++ Makefile
@@ -130,7 +130,7 @@ $(DIST_TARGET_ARCH_BIN): all $(DIST_DIR)/../.dummy $(LIBURB_TARGET) $(LIBURB_BIG
 	cp $(FETCHER_SRC) $(FETCHER_TARGET)
 	cp $(COMMAND_EXECUTOR_SRC) $(COMMAND_EXECUTOR_TARGET)
 	cd $(LIBURB_TARGET_DIR) && ln -sf $(LIB_NAME) $(LIB_SONAME) && ln -sf $(LIB_NAME) $(LIB_NAME_SHORT)
-	cd $(LIBURB_TARGET_DIR) && ln -sf $(LIB_BIG) $(LIB_SOBIG) && ln -sf $(LIB_BIG) $(LIB_BIG_SHORT)
+	if [ -f $(LIB_BIG) ]; then cd $(LIBURB_TARGET_DIR) && ln -sf $(LIB_BIG) $(LIB_SOBIG) && ln -sf $(LIB_BIG) $(LIB_BIG_SHORT); fi
 	cd $(BASE_DIR)/$(DEP_ARCH_BIN) && $(TAR) czvf $@ .
 
 $(DIST_TARGET_NOARCH): all $(DIST_DIR)/../.dummy $(URB_EGG_TARGET) $(LIBURB_PYTHON_NOARCH_BINDINGS)
@@ -247,6 +247,7 @@ ifdef FULL_MESOS_LIB
 $(LIBURB_BIG_TARGET): $(LIBURB_DIR)/$(LIB_BIG_TARGET)
 	cp -f $< $@
 endif
+
 $(URB_EGG_TARGET): $(URB_EGG_DIR)/dist/$(URB_EGG_NAME)
 	mkdir -p $(@D)
 	cp -f $< $@
